fix(oxygen): replace double with decimal for all monetary/rate values
ScoreTier.BaseRate, ScoreTier.MaxLoanPercentage, TermMultiplier.Value,
LoanApplication.Amount, ApplyForLoanRequest.LoanAmount, ILoanService
signature — all changed from double to decimal. Amortisation math
now uses DecimalPower.Pow (binary exponentiation) instead of Math.Pow
to avoid binary floating-point rounding errors.

+++ oxygen/Oxygen.Domain/ScoreTier.cs
@@ -3,10 +3,10 @@ namespace Oxygen.Domain;
 public class ScoreTier
 {
     public string Name { get; }
-    public double BaseRate { get; }
-    public double MaxLoanPercentage { get; }
+    public decimal BaseRate { get; }
+    public decimal MaxLoanPercentage { get; }
 
-    private ScoreTier(string name, double baseRate, double maxLoanPercentage)
+    private ScoreTier(string name, decimal baseRate, decimal maxLoanPercentage)
     {
         Name = name;
         BaseRate = baseRate;
@@ -16,15 +16,15 @@ public class ScoreTier
     public static ScoreTier For(int score)
     {
         if (score <= 599)
-            return new ScoreTier("Poor", 22, 0.15);
+            return new ScoreTier("Poor", 22m, 0.15m);
 
         return score switch
         {
-            <= 674 => new ScoreTier("Fair", 16, 0.25),
-            <= 724 => new ScoreTier("Good", 12, 0.35),
-            <= 774 => new ScoreTier("Very Good", 8, 0.45),
-            <= 849 => new ScoreTier("Excellent", 5, 0.55),
-            _ => new ScoreTier("Outstanding", 3, 0.65)
+            <= 674 => new ScoreTier("Fair", 16m, 0.25m),
+            <= 724 => new ScoreTier("Good", 12m, 0.35m),
+            <= 774 => new ScoreTier("Very Good", 8m, 0.45m),
+            <= 849 => new ScoreTier("Excellent", 5m, 0.55m),
+            _ => new ScoreTier("Outstanding", 3m, 0.65m)
         };
     }
 }
